requirements.txt
# FastAPI y servidor
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Base de datos
supabase==2.28.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# Autenticación y seguridad
python-dotenv==1.0.1

# Validación
pydantic>=2.11.7,<3.0.0
pydantic-settings==2.7.0
email-validator==2.2.0

# Utilidades
python-dateutil==2.9.0
pytz==2024.2

# PDF Generation
reportlab==4.2.5
weasyprint==63.1

# Email
aiosmtplib==3.0.2
jinja2==3.1.4

# File handling
python-magic==0.4.27

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0